move trade cutoff dates to utils
