+++ packages/utils/src/index.ts
@@ -14,6 +14,7 @@ export {
   isOptimismChainId
 } from './isL2ChainId';
 export { isCowSupportedChainId } from './isCowSupportedChainId';
+export { TRADE_CUTOFF_DATES } from './tradeCutoffDates';
 export { formatAddress } from './formatAddress';
 export { normalizeAddress } from './normalizeAddress';
 export { chainId } from './chainId';
